% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealr.R
\name{dealr}
\alias{dealr}
\title{Differential Expression Aware Ligand-Receptor inference}
\usage{
dealr(
  degList,
  db,
  baseMeanThresh = 100,
  abslogfcThresh = 1,
  baseMean_field = "baseMean",
  logfc_field = "log2FoldChange",
  stat_field = "stat",
  padj_field = "padj",
  pathway_name_field = "pathway_name",
  ligand_symbol_field = "ligand.symbol",
  receptor_symbol_field = "receptor.symbol",
  symbol_splitBy = ", "
)
}
\arguments{
\item{degList}{A named list of DESeq2 result data frames. See requirement in
detailed section below.}

\item{db}{A data frame containing ligand-receptor pair annotation. With each
row representing an LR-pair, it must have 1. unique rownames as for
identifying each LR-pair; 2. a column containing gene symbols
for the ligand; 3. a column containing gene symbols for the receptor; 4. a
column annotating the pathway the LR-pair belongs to. Complex molecules
transcribed from multiple genes should be presented with simple consistent
separator (e.g. 'Tgfbr1, Tgfbr2').}

\item{baseMeanThresh}{Numeric, threshold on the mean DESeq2-normalized
expression. Lower expression will be considered as not having sufficient
expression to be a ligand or receptor. Default \code{100}.}

\item{baseMean_field, stat_field, padj_field}{Column names to fetch necessary
statistics from each DE dataframe in \code{degList}. Default
\code{'baseMean'}, \code{'stat'}, and \code{'padj'}, respectively.}

\item{pathway_name_field, ligand_symbol_field, receptor_symbol_field}{Column
names to fetch necessary annotation from the LR-pair database \code{db}.
Default \code{'pathway_name'}, \code{'ligand.symbol'}, and
\code{'receptor.symbol'}, respectively.}

\item{symbol_splitBy}{The separator applied to process the ligand and
receptor symbols. Default \code{', '}.}
}
\value{
A \code{dealr} object, inherited from \code{tbl_df}, \code{dbl} and
\code{data.frame}. It contains the following columns:
\itemize{
\item{\code{LR_pair} - Identifier of the LR-pair, extracted from
\code{rownames(db)}}
\item{\code{sender} - Group name where the ligand statistics are from}
\item{\code{receiver} - Group name where the receptor statistics are from}
\item{\code{stat} - Combined z-score of the LR-pair between the sender and
receiver}
\item{\code{p} - P-value of the combined z-score}
\item{\code{pathway} - Pathway name of the LR-pair, extracted from
\code{db[[pathway_name_field]]}}
\item{\code{pathway_size} - Number of LR-pairs in the pathway, summarized
from \code{db} but not yielded from the analysis}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

DEALR infers differential ligand-receptor interactions from existing
differential gene expression (DGE) test results. It uses Stouffer's method to
combine the z-scores of ligands and receptors in pair to derive combined
z-scores and p-values. Equation below summarizes the z-score for an LR-pair
where ligand (\eqn{L}) is sent from group \eqn{i} and receptor (\eqn{R}) is
on group \eqn{j}. We sum up the z-score of \eqn{L} tested within group
\eqn{i} and the z-score of \eqn{R} tested within group \eqn{j} and divide by
the square root of the number of z-scores (A ligand or receptor can be a
complex of the product of multiple genes).

\deqn{
Z_{i,j,LR} = (\sum{Z_{L_i}} + \sum{Z_{R_j}})\sqrt{n}^{-1}
}

The p-value is then simply the probability that the combined z-score is
larger than the observed value, assuming mean 0 and standard deviation 1.

This method is assumes the DGE test to be performed with DESeq2 over
pseudo-bulk samples. Tests should be designed to compare to conditions
(e.g. treated vs control) within each group of cell (e.g. cell type).
Therefore, the input \code{degList} object is a list of DE result, each for
a group of cell.

The database tested to be useful came from CellChat.
}
\section{Expected input DE result}{

This tool currently only targets at scRNAseq dataset, with the study design
consisting of multiple conditions each having multiple replicates. The
prerequisite is the pseudo-bulk DESeq2 tests comparing two conditions within
each group of cell. For example,
\enumerate{
\item{Optionally, subset the dataset to only the two conditions of interests
if there are more;}
\item{Split the subset by cell type and do the following step for each cell
type;}
\item{Aggregate (sum up) raw counts by replicate. This builds the
pseudo-bulk. This naturally yields replicate-level metadata of for the
conditions of interests;}
\item{Run DESeq2 Wald test on the pseudo-bulk data with the design on the
condition.}
}

\strong{Make sure to have all tests contrasting the same direction}.

The procedure above should results in a number of DESeq2 result data frames.
This function expect a named list object that gathers all these data frames. List
names are for the cell types. \strong{Do NOT FILTER} the DESeq2 result data frames
before passing to this function.
}

\examples{
db_mini <- db$mouse[db$mouse$pathway_name \%in\% c('IL16', 'TNF'),]
lr <- dealr(deg_mini, db_mini)
}
