% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pathways}
\alias{pathways}
\alias{clusters}
\title{Access miscellaneous information in the results}
\usage{
pathways(x)

clusters(x)
}
\arguments{
\item{x}{A \code{dealr} or \code{dealr_pe} object.}
}
\value{
See Description.
}
\description{
\itemize{
\item \code{pathways()}: Returns a character vector of all unique pathways in the result.
\item \code{clusters()}: Returns a character vector of all unique clusters in the result.
}
}
