% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deg_mini}
\alias{deg_mini}
\title{Example DEG list input}
\format{
A list of three data.frames, named as \code{c('CD4_T', 'CD8_T', 'NK')}.
Each data.frame has the same 5 rows, corresponding to the 5 genes. Each
data.frame has the same columns and these are the exact output of DESeq2.
}
\usage{
deg_mini
}
\description{
This is a minimum subset of real analysis result from our in-house dataset.
Pseudo-bulk DESeq2 tests were performed between treated and control samples
within each cell type. This object contains the result for only three cell
types (CD4_T, CD8_T and NK), for minimum demonstration purpose. The DESeq2
result objects were coerced to plain \code{data.frame} and only contain test
statistics for 5 genes. These genes were chosen based on small pathways
(IL16 and TNF) that generally show significant intensity changes across the
three cell types.
}
\keyword{datasets}
