% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.dotplot}
\alias{.dotplot}
\title{Dot plot generalized function}
\usage{
.dotplot(
  plotDF,
  x,
  y,
  size,
  fill,
  downreg_col,
  upreg_col,
  fill_title,
  size_title,
  dot_size_range
)
}
\arguments{
\item{plotDF}{a data frame containing the data to be plotted.}

\item{x, y, size, fill}{Aesthetic mappings.}

\item{downreg_col}{a character string indicating the darkest color for
downregulated pathways. Default is \verb{#2166AC} (blue).}

\item{upreg_col}{a character string indicating the darkest color for
upregulated pathways. Default is \verb{#B2182B} (red).}

\item{fill_title}{a character string indicating the title of the fill
legend.}

\item{size_title}{a character string indicating the title of the size
legend.}

\item{dot_size_range}{a numeric vector of length 2 indicating the range of
dot sizes. The first value is the minimum size, and the second value is the
maximum size. Default is \code{c(0.3, 4)}.}
}
\value{
ggplot object to be further customized.
}
\description{
Dot plot generalized function
}
