% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.theme_text_setter}
\alias{.theme_text_setter}
\title{Text size setter}
\usage{
.theme_text_setter(
  text_x_size = 8,
  text_y_size = 6,
  text_strip_size = 6,
  text_title_size = 10,
  text_legend_size = 8,
  text_legend_title_size = 8
)
}
\arguments{
\item{text_x_size}{Size of cluster labels on x-axis. Default \code{8}.}

\item{text_y_size}{Size of LR-pair or pathway labels on y-axis. Default \code{6}.}

\item{text_strip_size}{Size of pathway labels in the facet strip. Default \code{6}.}

\item{text_title_size}{Size of plot title. Default \code{10}.}

\item{text_legend_size}{Size of text in legend items. Default \code{8}.}

\item{text_legend_title_size}{Size of legend title. Default \code{8}.}
}
\description{
Text size setter
}
