% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealr.R
\name{lookback_input}
\alias{lookback_input}
\title{Helper func: Look back to input stats}
\usage{
lookback_input(
  degList,
  db,
  pathway_use,
  sender_use,
  receiver_use,
  pathway_name_field = "pathway_name",
  ligand_symbol_field = "ligand.symbol",
  receptor_symbol_field = "receptor.symbol",
  symbol_splitBy = ", "
)
}
\arguments{
\item{degList}{The \code{degList} used by \code{\link[=dealr]{dealr()}}.}

\item{db}{The \code{db} used by \code{\link[=dealr]{dealr()}}.}

\item{pathway_use}{The name of the pathway to look at.}

\item{sender_use}{The name of the sender group.}

\item{receiver_use}{The name of the receiver group.}

\item{pathway_name_field, ligand_symbol_field, receptor_symbol_field}{Column
names to fetch necessary annotation from the LR-pair database \code{db}.
Default \code{'pathway_name'}, \code{'ligand.symbol'}, and
\code{'receptor.symbol'}, respectively.}

\item{symbol_splitBy}{The separator applied to process the ligand and
receptor symbols. Default \code{', '}.}
}
\value{
No return value. Print information to the console.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

A helper function that might be removed in the future, mainly for debugging.
It takes the input \code{degList} and \code{db} and show the original input statistics
for the ligands and receptors of the LR-pairs in the pathway. Just to see if
we are making good decision.

Basically what you want to see: 1. If the \code{baseMean} is high enough to
say the component is stably expressed, regardless of whether there is
significant regulation. Low \code{lfcSE} also helps indicating whether the
expression is stable. 2. If \code{padj} is NA, it means DESeq2 already
ignored it due to too-low expression. 3. Then see \code{stat} for verifying
if the calculation is correct.
}
