% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSenderLR}
\alias{plotSenderLR}
\alias{plotReceiverLR}
\title{Dot plot showing top differential signaling LR-pairs}
\usage{
plotSenderLR(
  dealr,
  sender_use,
  p_thresh = 0.01,
  pathway_use = NULL,
  top_n = NULL,
  downreg_col = "#2166AC",
  upreg_col = "#B2182B",
  dot_size_range = c(0.3, 4)
)

plotReceiverLR(
  dealr,
  receiver_use,
  p_thresh = 0.01,
  pathway_use = NULL,
  top_n = NULL,
  downreg_col = "#2166AC",
  upreg_col = "#B2182B",
  dot_size_range = c(0.3, 4)
)
}
\arguments{
\item{dealr}{a \code{dealr} object. DO NOT FILTER.}

\item{sender_use}{a character string indicating the sender to be plotted.}

\item{p_thresh}{a numeric value indicating the p-value threshold for
filtering the LR-pairs. Default is \code{0.01}.}

\item{pathway_use}{Character vector of pathways to filter the LR-pairs.
Applied after p-value filtering but prior to top n slicing. Default
\code{NULL}.}

\item{top_n}{a numeric value indicating the number of top LR-pairs to be
plotted, ranked by p-value. Default \code{NULL} means all LR-pairs, usually
too many.}

\item{downreg_col}{a character string indicating the darkest color for
downregulated LR-pairs. Default is \verb{#2166AC} (blue).}

\item{upreg_col}{a character string indicating the darkest color for
upregulated LR-pairs. Default is \verb{#B2182B} (red).}

\item{dot_size_range}{a numeric vector of length 2 indicating the range of
dot sizes. The first value is the minimum size, and the second value is the
maximum size. Default is \code{c(0.3, 4)}.}

\item{receiver_use}{a character string indicating the receiver to be plotted.}
}
\value{
ggplot object of the dot plot.
}
\description{
\itemize{
\item{\code{plotSenderLR} - Focuses on one sender on top, between any
receiver groups at bottom}
\item{\code{plotReceiverLR} - Focuses on one receiver at the bottom, between
any sender groups on the top}
}

The size of dots reflects significance of the LR-pairs, and the color
indicates the direction of the interaction.
}
