% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPathwayEnrichDot}
\alias{plotPathwayEnrichDot}
\title{Dot plot showing pathway enrichment from differential signaling LR-pairs}
\usage{
plotPathwayEnrichDot(
  dealr_pe,
  sender_use = NULL,
  receiver_use = NULL,
  focus = NULL,
  p_thresh = 0.01,
  size_by = c("overlap", "enrichment"),
  fe_thresh = 2,
  top_n = NULL,
  downreg_col = "#2166AC",
  downreg_name = "Control",
  upreg_col = "#B2182B",
  upreg_name = "Test",
  dot_size_range = c(0.3, 4),
  text_x_size = 8,
  text_y_size = 6,
  text_title_size = 10,
  text_legend_size = 8,
  text_legend_title_size = 8,
  text_strip_size = 6
)
}
\arguments{
\item{dealr_pe}{a \code{dealr_pe} object. DO NOT FILTER.}

\item{sender_use}{Character vector of sender groups to be considered. Default
\code{NULL} includes all possible senders.}

\item{receiver_use}{Character vector of receiver groups to be considered.
Default \code{NULL} includes all possible receivers.}

\item{focus}{A single character string indicating a group of interest. If
provided, only show the communication involving this group and the plot
will be split to two panels for when this group is sending the ligand and
when this group is receiving the ligand. Default \code{NULL} use all possible
pairs of sender and receiver as provided with \code{sender_use} and
\code{receiver_use}.}

\item{p_thresh}{a numeric value indicating the p-value threshold for
filtering the pathways. Default \code{0.01}.}

\item{size_by}{Choose from \code{'overlap'} or \code{'enrichment'}. If \code{overlap}, dots
are sized by the number of significant LR-pairs annotated as in the pathway.
If \code{enrichment}, dots are sized by the log of the fold enrichment score.
Default \code{overlap}.}

\item{fe_thresh}{a numeric value indicating the fold enrichment threshold
for filtering the pathways. Default \code{2}.}

\item{top_n}{a numeric value indicating the number of top pathways to be
plotted, ranked by p-value. Default \code{NULL} means all pathways, usually
not too many.}

\item{downreg_col}{a character string indicating the darkest color for
downregulated pathways. Default is \verb{#2166AC} (blue).}

\item{upreg_col}{a character string indicating the darkest color for
upregulated pathways. Default is \verb{#B2182B} (red).}

\item{dot_size_range}{a numeric vector of length 2 indicating the range of
dot sizes. The first value is the minimum size, and the second value is the
maximum size. Default is \code{c(0.3, 4)}.}

\item{text_x_size}{Size of cluster labels on x-axis. Default \code{8}.}

\item{text_y_size}{Size of LR-pair or pathway labels on y-axis. Default \code{6}.}

\item{text_title_size}{Size of plot title. Default \code{10}.}

\item{text_legend_size}{Size of text in legend items. Default \code{8}.}

\item{text_legend_title_size}{Size of legend title. Default \code{8}.}

\item{text_strip_size}{Size of pathway labels in the facet strip. Default \code{6}.}
}
\value{
ggplot object of the dot plot.
}
\description{
The size of dots shows the fold enrichment of the pathways, and the color
indicates the significance and direction of the regulation of the pathways.
}
\examples{
db_mini <- db$mouse[db$mouse$pathway_name \%in\% c('IL16', 'TNF'),]
lr <- dealr(deg_mini, db_mini)
pe <- pathwayEnrich(lr)
plotPathwayEnrichDot(pe)
}
