% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotLRPairDot}
\alias{plotLRPairDot}
\title{Dot plot showing top differential signaling LR-pairs}
\usage{
plotLRPairDot(
  dealr,
  sender_use = NULL,
  receiver_use = NULL,
  focus = NULL,
  pathway_use = NULL,
  p_thresh = 0.01,
  top_n = NULL,
  downreg_col = "#2166AC",
  upreg_col = "#B2182B",
  dot_size_range = c(0.3, 4),
  text_x_size = 8,
  text_y_size = 6,
  text_strip_size = 6,
  text_title_size = 10,
  text_legend_size = 8,
  text_legend_title_size = 8
)
}
\arguments{
\item{dealr}{a \code{dealr} object. DO NOT FILTER.}

\item{sender_use}{Character vector of sender groups to be considered. Default
\code{NULL} includes all possible senders.}

\item{receiver_use}{Character vector of receiver groups to be considered.
Default \code{NULL} includes all possible receivers.}

\item{focus}{A single character string indicating a group of interest. If
provided, only show the communication involving this group and the plot
will be split to two panels for when this group is sending the ligand and
when this group is receiving the ligand. Default \code{NULL} use all possible
pairs of sender and receiver as provided with \code{sender_use} and
\code{receiver_use}.}

\item{pathway_use}{Character vector of pathways to filter the LR-pairs.
Applied after p-value filtering but prior to top n slicing. Default
\code{NULL}.}

\item{p_thresh}{a numeric value indicating the p-value threshold for
filtering the LR-pairs. Default is \code{0.01}.}

\item{top_n}{a numeric value indicating the number of top LR-pairs to be
plotted, ranked by p-value. Default \code{NULL} means all LR-pairs, usually
too many.}

\item{downreg_col}{a character string indicating the darkest color for
downregulated pathways. Default is \verb{#2166AC} (blue).}

\item{upreg_col}{a character string indicating the darkest color for
upregulated pathways. Default is \verb{#B2182B} (red).}

\item{dot_size_range}{a numeric vector of length 2 indicating the range of
dot sizes. The first value is the minimum size, and the second value is the
maximum size. Default is \code{c(0.3, 4)}.}

\item{text_x_size}{Size of cluster labels on x-axis. Default \code{8}.}

\item{text_y_size}{Size of LR-pair or pathway labels on y-axis. Default \code{6}.}

\item{text_strip_size}{Size of pathway labels in the facet strip. Default \code{6}.}

\item{text_title_size}{Size of plot title. Default \code{10}.}

\item{text_legend_size}{Size of text in legend items. Default \code{8}.}

\item{text_legend_title_size}{Size of legend title. Default \code{8}.}
}
\value{
ggplot object of the dot plot.
}
\description{
The size of dots reflects significance of the LR-pairs, and the color
indicates the direction of the interaction.
}
\examples{
db_mini <- db$mouse[db$mouse$pathway_name \%in\% c('IL16', 'TNF'),]
lr <- dealr(deg_mini, db_mini)
plotLRPairDot(lr)
}
