% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealr.R
\name{print.dealr}
\alias{print.dealr}
\title{Show collapsed information of DEALR result}
\usage{
\method{print}{dealr}(x, trunc = 10, ...)
}
\arguments{
\item{x}{A \code{dealr} object.}

\item{trunc}{Number of maximum list element to show.}

\item{...}{Additional arguments (not used).}
}
\description{
Show collapsed information of DEALR result
}
