% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealr.R
\name{print.dealr_pe}
\alias{print.dealr_pe}
\title{Show collapsed information of DEALR Pathway Enrichment result}
\usage{
\method{print}{dealr_pe}(x, trunc = 10, ...)
}
\arguments{
\item{x}{A \code{dealr_pe} object.}

\item{trunc}{Number of maximum list element to show.}

\item{...}{Additional arguments (not used).}
}
\description{
Show collapsed information of DEALR Pathway Enrichment result
}
\examples{
db_mini <- db$mouse[db$mouse$pathway_name \%in\% c('IL16', 'TNF'),]
lr <- dealr(deg_mini, db_mini)
pe <- pathwayEnrich(lr)
print(pe)
}
