% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{pseudobulkDE}
\alias{pseudobulkDE}
\alias{pseudobulkDE.default}
\alias{pseudobulkDE.liger}
\alias{pseudobulkDE.Seurat}
\title{Perform pseudo-bulk differential expression test}
\usage{
pseudobulkDE(
  data,
  condVar,
  condTest,
  condCtrl,
  splitVar,
  replicateVar = NULL,
  ...
)

\method{pseudobulkDE}{default}(
  data,
  condVar,
  condTest,
  condCtrl,
  splitVar,
  replicateVar = NULL,
  returnDDS = FALSE,
  verbose = TRUE,
  ...
)

\method{pseudobulkDE}{liger}(
  data,
  condVar,
  condTest,
  condCtrl,
  splitVar = NULL,
  replicateVar = NULL,
  returnDDS = FALSE,
  verbose = TRUE,
  ...
)

\method{pseudobulkDE}{Seurat}(
  data,
  condVar,
  condTest,
  condCtrl,
  splitVar = NULL,
  replicateVar = NULL,
  returnDDS = FALSE,
  verbose = TRUE,
  assay = "RNA",
  ...
)
}
\arguments{
\item{data}{Full raw counts expression data of the single cells. Can be a
single matrix or a list of matrices without merging, a commonly seen
container object with raw counts at its conventional slot (e.g.
LayerData(seurat, 'counts'), counts(sce), or rawData(liger)).}

\item{condVar}{The condition variable that matches to all cells in the data.}

\item{condTest}{The test condition within the condition variable.}

\item{condCtrl}{The control condition within the condition variable.}

\item{splitVar}{The variable that splits the data into different identity
classes.}

\item{replicateVar}{The variable indicating the biological replicate. Default
\code{NULL}.}

\item{...}{Arguments passed to methods and \code{DESeq2::DESeq()}. Note that
\code{quite = TRUE} is pre-occupied.}

\item{returnDDS}{Logical, whether to return the DESeqDataSet object instead
of just the result statistics. Default \code{FALSE}.}

\item{verbose}{Logical, whether to print some useful information. Default
\code{TRUE}}

\item{assay}{Name of assay to fetch counts from a Seurat object. Default
\code{"RNA"}.}
}
\value{
A list of DESeq2 result objects, one for each identity class.
}
\description{
This function performs pseudo-bulk differential expression test comparing
two conditions (e.g. treated vs. control) within each identity class (e.g.
cell type or cluster). For within each identity class, we group the cells by
condition and biological replicate variable, and then sum the counts to
generate pseudo-bulk counts. The pseudo-bulk counts are then sent to DESeq2
for differential expression analysis.
}
\details{
For the input of a plain matrix or a list of matrices, any matrix coercible
to \code{dgCMatrix} is accepted. Arguments \code{condVar}, \code{splitVar} and
\code{replicateVar} must be a factor object matching all cells. For a single
matrix, these variables must be in the same order as the columns of the
matrix. For a list of matrices, the order follows as if the matrices inside
are concatenated in order. When presenting a list of matrices, \code{replicateVar}
is by default automatically generated following the matrix belonging.

For container objects, \code{condVar}, \code{splitVar} and \code{replicateVar} are then
expected to be variable names in the metadata slot. For classes with
conventional fields in metadata, defaults are taken. For example, for a
Seurat object, \code{splitVar} is by default \code{Idents(data)} and \code{replicateVar} is
by default \code{orig.ident}. For a \code{liger} object, \code{splitVar} is by default
\code{defaultCluster} and \code{replicateVar} is \code{data$dataset}.
}
