% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealr.R
\name{stouffer_z}
\alias{stouffer_z}
\title{Stouffer's method for combining z-scores}
\usage{
stouffer_z(z)
}
\arguments{
\item{z}{Numeric vector of z-scores.}
}
\value{
A numeric value representing the combined z-score.
}
\description{
Stouffer's method for combining z-scores
}
\details{
NA input values are replaced with 0.
}
\examples{
z <- c(1.5, 2.0, NA, 3.0)
stouffer_z(z)
}
