% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealr-package.R
\docType{package}
\name{dealr-package}
\alias{dealr-package}
\title{dealr: Differentially Expression Aware Ligand-Receptor Inference}
\description{
Inference differential ligand-receptor signaling based on differential gene expression statistics.
}
\author{
\strong{Maintainer}: Yichen Wang \email{wayichen@umich.edu}

Other contributors:
\itemize{
  \item Joshua Welch \email{welchjd@umich.edu} [copyright holder]
}

}
\keyword{internal}
