% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSenderPE}
\alias{plotSenderPE}
\alias{plotReceiverPE}
\title{Dot plot showing pathway enrichment from differential signaling LR-pairs}
\usage{
plotSenderPE(
  dealr_pe,
  sender_use,
  p_thresh = 0.01,
  top_n = NULL,
  downreg_col = "#2166AC",
  upreg_col = "#B2182B",
  dot_size_range = c(0.3, 4)
)

plotReceiverPE(
  dealr_pe,
  receiver_use,
  p_thresh = 0.01,
  top_n = NULL,
  downreg_col = "#2166AC",
  upreg_col = "#B2182B",
  dot_size_range = c(0.3, 4)
)
}
\arguments{
\item{dealr_pe}{a \code{dealr_pe} object. DO NOT FILTER.}

\item{sender_use}{a character string indicating the sender to be plotted.}

\item{p_thresh}{a numeric value indicating the p-value threshold for
filtering the pathways. Default \code{0.01}.}

\item{top_n}{a numeric value indicating the number of top pathways to be
plotted, ranked by p-value. Default \code{NULL} means all pathways, usually
not too many.}

\item{downreg_col}{a character string indicating the darkest color for
downregulated pathways. Default is \verb{#2166AC} (blue).}

\item{upreg_col}{a character string indicating the darkest color for
upregulated pathways. Default is \verb{#B2182B} (red).}

\item{dot_size_range}{a numeric vector of length 2 indicating the range of
dot sizes. The first value is the minimum size, and the second value is the
maximum size. Default is \code{c(0.3, 4)}.}

\item{receiver_use}{a character string indicating the receiver to be plotted.}
}
\value{
ggplot object of the dot plot.
}
\description{
\itemize{
\item{\code{plotSenderPE} - Focuses on one sender on top, between any
receiver groups at bottom}
\item{\code{plotReceiverPE} - Focuses on one receiver at the bottom, between
any sender groups on the top}
}
The size of dots shows the fold enrichment of the pathways, and the color
indicates the significance and direction of the regulation of the pathways.
}
