% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotLRGeneHeatmap}
\alias{plotLRGeneHeatmap}
\alias{plotLRGeneHeatmap.default}
\title{Gene expression heatmap of differential interacting LR-pair}
\usage{
plotLRGeneHeatmap(
  data,
  dealr,
  splitVar,
  sender_use,
  receiver_use,
  condVar,
  condTest,
  condCtrl,
  pval_thresh = 0.01,
  abslogfc_thresh = 1,
  top_n = NULL,
  pathway_use = NULL,
  splitVar_col = NULL,
  condVar_col = NULL,
  text_gene_size = 6,
  text_pathway_size = 8,
  text_condition_size = 10,
  text_annotation_size = 8,
  text_legend_size = 8,
  text_legend_title_size = 10,
  left_width = grid::unit(4, "cm"),
  middle_width = grid::unit(1, "cm"),
  right_width = grid::unit(4, "cm"),
  ...
)

\method{plotLRGeneHeatmap}{default}(
  data,
  dealr,
  splitVar,
  sender_use,
  receiver_use,
  condVar,
  condTest,
  condCtrl,
  pval_thresh = 0.01,
  abslogfc_thresh = 1,
  top_n = NULL,
  pathway_use = NULL,
  splitVar_col = NULL,
  condVar_col = NULL,
  text_gene_size = 6,
  text_pathway_size = 8,
  text_condition_size = 10,
  text_annotation_size = 8,
  text_legend_size = 8,
  text_legend_title_size = 10,
  left_width = grid::unit(4, "cm"),
  middle_width = grid::unit(1, "cm"),
  right_width = grid::unit(4, "cm"),
  ...
)
}
\description{
It creates heatmap focusing on a specific pair of sender and receiver cell
groups. The heatmap is further divided to show the contrast of between the
conditions where the input DEGs were derived. The columns of the heatmap
represent cells and each row is a gene. Rows are grouped by LR-pairs in the
situation when complex components are involved, and are further grouped by
pathways.
}
